0.1.1
Added basic SSH functionality along with X11 forwarding. Currently, the 'Run Script' button will now provide basic information about the VM. This is performed using the ganymed library (https://github.com/hudson/ganymed-ssh-2).

Currently the IP address, port, username and password are hard-coded to fit the primary VM I am working with. In the next update, the user will be allowed to set these fields in the form.

Extra functionality to be added next includes expanding the SSH functionality by adding SFTP to transfer files to the VM, and opening a shell on the VM to allow the user to input commands.